Add Cloverleaf (#269)
**https://cloverleaf.app**

**An open source app to replace your password manager without storing your passwords anywhere. **

Cloverleaf is a web app that I've been working on for 4 years now with the hope of making using secure passwords easy and painless. All you need to do is enter the name of the app you want to make a password for and your super secret master password. This'll then give you a password that fits all that site's requirements without storing it anywhere. 

# By submitting this pull request I confirm I've read and complied with the below requirements.

Failure to properly do so will just result in the pull request being closed and everyone's time wasted. Please read it twice. Most people miss many things.

- [x] I have read the [Contribution guidelines](https://github.com/aviaryan/awesome-no-login-web-apps/blob/master/CONTRIBUTING.md) and I am confident that my PR is suitable. 
- [x] This pull request has a descriptive title. For example, `Add [App name]`, not `Update readme.md` or `Added new entries`.
- [x] The app I added is not a duplicate.